# Conflicts:
#	core/src/com/mygdx/shapewars/controller/ShapeWarsController.java
#	core/src/com/mygdx/shapewars/model/components/VelocityComponent.java

+++ core/src/com/mygdx/shapewars/model/components/VelocityComponent.java
@@ -3,36 +3,36 @@ package com.mygdx.shapewars.model.components;
 import com.badlogic.ashley.core.Component;
 
 public class VelocityComponent implements Component{
-  private int value;
-  private int direction;
+  private float value;
+  private float direction;
 
   public VelocityComponent(int value, int direction) {
     this.value = value;
     this.direction = direction;
   }
 
-  public int getValue() {
+  public float getValue() {
     return value;
   }
 
-  public int getDirection() {
+  public float getDirection() {
     return direction;
   }
 
-  public void setVelocity(int v, int d) {
+  public void setVelocity(float v, float d) {
     value = v;
     direction = d;
   }
 
-  public void setValue(int v) {
+  public void setValue(float v) {
     setVelocity(v, direction);
   }
 
-  public void setDirection(int d) {
+  public void setDirection(float d) {
     setVelocity(value, d);
   }
 
-  public void addDirection(int d) {
+  public void addDirection(float d) {
     direction += d;
   }
 }
